Fixed (hack) replicated actor visibility when placed directly in room levels
This is a warkaround to avoid replicated actor's visibility messed up when they are placed directly in room levels.
ActorHiddenInGame is replicated and thus when a server RoomLevel toggles the visibility of a replicated actor, it will changes the visibility also on clients.
So, I excluded the replicated actors from the RoomLevel visibility toggle.
Instead, a RoomVisibilityComponent should be added to the replicated actors to toggle their visibility without any issues.